+++ docs/search/all_a.js
@@ -1,6 +1,8 @@
 var searchData=
 [
   ['id',['id',['../interfaceBristolSU_1_1Support_1_1ModuleInstance_1_1Contracts_1_1ModuleInstance.html#a4bbb1a3313e8223a9b4cf8713ae024b7',1,'BristolSU\Support\ModuleInstance\Contracts\ModuleInstance\id()'],['../classBristolSU_1_1Support_1_1ModuleInstance_1_1ModuleInstance.html#a8b009e11aeac0ba84f57a6dafae96d47',1,'BristolSU\Support\ModuleInstance\ModuleInstance\id()']]],
+  ['incorrectlogin',['IncorrectLogin',['../classBristolSU_1_1Support_1_1Authorization_1_1Exception_1_1IncorrectLogin.html',1,'BristolSU::Support::Authorization::Exception']]],
+  ['incorrectlogin_2ephp',['IncorrectLogin.php',['../IncorrectLogin_8php.html',1,'']]],
   ['index',['index',['../classBristolSU_1_1Support_1_1Http_1_1Controllers_1_1ModuleInstanceRedirectController.html#a1fb141246dd8360e3d9f4cfb7565dbdc',1,'BristolSU::Support::Http::Controllers::ModuleInstanceRedirectController']]],
   ['initializehasrevisions',['initializeHasRevisions',['../classBristolSU_1_1Support_1_1Revision_1_1HasRevisions.html#a853f86eacac35f26f6c84b9aa7bc1935',1,'BristolSU::Support::Revision::HasRevisions']]],
   ['injectactivity',['InjectActivity',['../classBristolSU_1_1Support_1_1Activity_1_1Middleware_1_1InjectActivity.html',1,'BristolSU::Support::Activity::Middleware']]],