+++ docs/search/all_b.js
@@ -1,13 +1,13 @@
 var searchData=
 [
-  ['laravelauthactivityinstanceresolver',['LaravelAuthActivityInstanceResolver',['../classBristolSU_1_1Support_1_1ActivityInstance_1_1LaravelAuthActivityInstanceResolver.html',1,'BristolSU::Support::ActivityInstance']]],
+  ['laravelauthactivityinstanceresolver',['LaravelAuthActivityInstanceResolver',['../classBristolSU_1_1Support_1_1Testing_1_1ActivityInstance_1_1LaravelAuthActivityInstanceResolver.html',1,'BristolSU::Support::Testing::ActivityInstance']]],
   ['laravelauthactivityinstanceresolver_2ephp',['LaravelAuthActivityInstanceResolver.php',['../LaravelAuthActivityInstanceResolver_8php.html',1,'']]],
   ['loadmigrations',['loadMigrations',['../classBristolSU_1_1Support_1_1Module_1_1ModuleServiceProvider.html#a0b4cc1cd93a8b14f435f426070d38a94',1,'BristolSU::Support::Module::ModuleServiceProvider']]],
   ['log',['Log',['../classBristolSU_1_1Support_1_1Action_1_1Actions_1_1Log.html',1,'BristolSU::Support::Action::Actions']]],
   ['log_2ephp',['Log.php',['../Log_8php.html',1,'']]],
   ['logic',['Logic',['../classBristolSU_1_1Support_1_1Logic_1_1Logic.html',1,'BristolSU\Support\Logic\Logic'],['../classBristolSU_1_1Support_1_1Filters_1_1FilterInstance.html#a9a3d4f0fc68ddf32d6a0f0a282b183c5',1,'BristolSU\Support\Filters\FilterInstance\logic()'],['../classBristolSU_1_1Support_1_1Permissions_1_1Models_1_1ModuleInstancePermission.html#a08fc219c42e6ae8011af2dd7208cffd0',1,'BristolSU\Support\Permissions\Models\ModuleInstancePermission\logic()']]],
   ['logic_2ephp',['Logic.php',['../Logic_8php.html',1,'']]],
-  ['logicaudience',['LogicAudience',['../classBristolSU_1_1Support_1_1Logic_1_1Contracts_1_1Audience_1_1LogicAudience.html',1,'BristolSU\Support\Logic\Contracts\Audience\LogicAudience'],['../classBristolSU_1_1Support_1_1Logic_1_1Audience_1_1LogicAudience.html',1,'BristolSU\Support\Logic\Audience\LogicAudience']]],
+  ['logicaudience',['LogicAudience',['../classBristolSU_1_1Support_1_1Logic_1_1Audience_1_1LogicAudience.html',1,'BristolSU\Support\Logic\Audience\LogicAudience'],['../classBristolSU_1_1Support_1_1Logic_1_1Contracts_1_1Audience_1_1LogicAudience.html',1,'BristolSU\Support\Logic\Contracts\Audience\LogicAudience']]],
   ['logicaudience_2ephp',['LogicAudience.php',['../Audience_2LogicAudience_8php.html',1,'(Global Namespace)'],['../Contracts_2Audience_2LogicAudience_8php.html',1,'(Global Namespace)']]],
   ['logicrepository',['LogicRepository',['../interfaceBristolSU_1_1Support_1_1Logic_1_1Contracts_1_1LogicRepository.html',1,'BristolSU\Support\Logic\Contracts\LogicRepository'],['../classBristolSU_1_1Support_1_1Logic_1_1LogicRepository.html',1,'BristolSU\Support\Logic\LogicRepository']]],
   ['logicrepository_2ephp',['LogicRepository.php',['../Contracts_2LogicRepository_8php.html',1,'(Global Namespace)'],['../LogicRepository_8php.html',1,'(Global Namespace)']]],
@@ -19,8 +19,6 @@ var searchData=
   ['logictesterfake_2ephp',['LogicTesterFake.php',['../LogicTesterFake_8php.html',1,'']]],
   ['logictesterresult',['LogicTesterResult',['../classBristolSU_1_1Support_1_1Testing_1_1LogicTester_1_1LogicTesterResult.html',1,'BristolSU::Support::Testing::LogicTester']]],
   ['logictesterresult_2ephp',['LogicTesterResult.php',['../LogicTesterResult_8php.html',1,'']]],
-  ['logintoactivityinstance',['LogIntoActivityInstance',['../classBristolSU_1_1Support_1_1ActivityInstance_1_1Middleware_1_1LogIntoActivityInstance.html',1,'BristolSU::Support::ActivityInstance::Middleware']]],
-  ['logintoactivityinstance_2ephp',['LogIntoActivityInstance.php',['../LogIntoActivityInstance_8php.html',1,'']]],
   ['logout',['logout',['../interfaceBristolSU_1_1Support_1_1User_1_1Contracts_1_1UserAuthentication.html#a75f8ce299a4d8f3d4d7d1ac6773919a3',1,'BristolSU\Support\User\Contracts\UserAuthentication\logout()'],['../classBristolSU_1_1Support_1_1User_1_1UserApiAuthentication.html#a024516f9cec5aa8106dbdc1ee87e8768',1,'BristolSU\Support\User\UserApiAuthentication\logout()'],['../classBristolSU_1_1Support_1_1User_1_1UserWebAuthentication.html#a127a6e239ae326d1447daf619f590c8f',1,'BristolSU\Support\User\UserWebAuthentication\logout()']]],
   ['logoutofextras',['LogoutOfExtras',['../classBristolSU_1_1Support_1_1Authorization_1_1Middleware_1_1LogoutOfExtras.html',1,'BristolSU::Support::Authorization::Middleware']]],
   ['logoutofextras_2ephp',['LogoutOfExtras.php',['../LogoutOfExtras_8php.html',1,'']]]