+++ docs/search/classes_11.js
@@ -1,4 +1,5 @@
 var searchData=
 [
-  ['websessionauthentication',['WebSessionAuthentication',['../classBristolSU_1_1Support_1_1Authentication_1_1WebSessionAuthentication.html',1,'BristolSU::Support::Authentication']]]
+  ['webrequestactivityinstanceresolver',['WebRequestActivityInstanceResolver',['../classBristolSU_1_1Support_1_1ActivityInstance_1_1WebRequestActivityInstanceResolver.html',1,'BristolSU::Support::ActivityInstance']]],
+  ['webrequestauthentication',['WebRequestAuthentication',['../classBristolSU_1_1Support_1_1Authentication_1_1WebRequestAuthentication.html',1,'BristolSU::Support::Authentication']]]
 ];