+++ docs/search/classes_7.js
@@ -1,5 +1,6 @@
 var searchData=
 [
+  ['incorrectlogin',['IncorrectLogin',['../classBristolSU_1_1Support_1_1Authorization_1_1Exception_1_1IncorrectLogin.html',1,'BristolSU::Support::Authorization::Exception']]],
   ['injectactivity',['InjectActivity',['../classBristolSU_1_1Support_1_1Activity_1_1Middleware_1_1InjectActivity.html',1,'BristolSU::Support::Activity::Middleware']]],
   ['injectactivityinstance',['InjectActivityInstance',['../classBristolSU_1_1Support_1_1ActivityInstance_1_1Middleware_1_1InjectActivityInstance.html',1,'BristolSU::Support::ActivityInstance::Middleware']]],
   ['injectjavascriptvariables',['InjectJavascriptVariables',['../classBristolSU_1_1Support_1_1Http_1_1Middleware_1_1InjectJavascriptVariables.html',1,'BristolSU::Support::Http::Middleware']]],