+++ tests/Http/Middleware/InjectJavascriptVariablesTest.php
@@ -4,6 +4,7 @@ namespace BristolSU\Support\Tests\Http\Middleware;
 
 use BristolSU\Support\ActivityInstance\ActivityInstance;
 use BristolSU\Support\ActivityInstance\Contracts\ActivityInstanceResolver;
+use BristolSU\Support\ActivityInstance\Exceptions\NotInActivityInstanceException;
 use BristolSU\Support\Authentication\Contracts\Authentication;
 use BristolSU\Support\Http\Middleware\InjectJavascriptVariables;
 use BristolSU\Support\Testing\CreatesModuleEnvironment;
@@ -179,4 +180,53 @@ class InjectJavascriptVariablesTest extends TestCase
         });
     }
 
+    /** @test */
+    public function it_sets_the_activity_instance()
+    {
+        $request = $this->prophesize(Request::class);
+        $request->route('module_instance_slug')->shouldBeCalled()->willReturn($this->getModuleInstance());
+        $request->route('activity_slug')->shouldBeCalled()->willReturn($this->getActivity());
+        $request->is('a/*')->shouldBeCalled()->willReturn(false);
+        
+        $activityInstance = factory(ActivityInstance::class)->create();
+        
+        $activityInstanceResolver = $this->prophesize(ActivityInstanceResolver::class);
+        $activityInstanceResolver->getActivityInstance()->shouldBeCalled()->willReturn($activityInstance);
+        $this->instance(ActivityInstanceResolver::class, $activityInstanceResolver->reveal());
+        
+        $javascript = $this->prophesize(Transformer::class);
+        $javascript->put(Argument::that(function ($arg) use ($activityInstance) {
+            return array_key_exists('activityinstance', $arg) && $arg['activityinstance'] instanceof ActivityInstance
+                && $arg['activityinstance']->is($activityInstance);
+        }))->shouldBeCalled();
+        $this->instance('JavaScript', $javascript->reveal());
+
+        $injector = resolve(InjectJavascriptVariables::class);
+        $injector->handle($request->reveal(), function ($request) {
+        });
+    }
+
+
+    /** @test */
+    public function it_sets_the_activity_instance_to_null_if_no_activity_instance_found()
+    {
+        $request = $this->prophesize(Request::class);
+        $request->route('module_instance_slug')->shouldBeCalled()->willReturn($this->getModuleInstance());
+        $request->route('activity_slug')->shouldBeCalled()->willReturn($this->getActivity());
+        $request->is('a/*')->shouldBeCalled()->willReturn(false);
+        
+        $activityInstanceResolver = $this->prophesize(ActivityInstanceResolver::class);
+        $activityInstanceResolver->getActivityInstance()->shouldBeCalled()->willThrow(new NotInActivityInstanceException());
+        $this->instance(ActivityInstanceResolver::class, $activityInstanceResolver->reveal());
+
+        $javascript = $this->prophesize(Transformer::class);
+        $javascript->put(Argument::that(function ($arg) {
+            return array_key_exists('activityinstance', $arg) && $arg['activityinstance'] === null;
+        }))->shouldBeCalled();
+        $this->instance('JavaScript', $javascript->reveal());
+
+        $injector = resolve(InjectJavascriptVariables::class);
+        $injector->handle($request->reveal(), function ($request) {
+        });
+    }
 }